static_call/x86: Add __static_call_return0()
commit 3f2a8fc4b15de18644e8a80a09edda168676e22c  upstream

Provide a stub function that return 0 and wire up the static call site
patching to replace the CALL with a single 5 byte instruction that
clears %RAX, the return value register.

The function can be cast to any function pointer type that has a
single %RAX return (including pointers). Also provide a version that
returns an int for convenience. We are clearing the entire %RAX register
in any case, whether the return value is 32 or 64 bits, since %RAX is
always a scratch register anyway.

[Backport changes]

1.In include/linux/static_call.h, DEFINE_STATIC_CALL() is changed
to _DEFINE_STATIC_CALL() with three arguments to support the RET0 call
implementation. This is picked from 29fd01944b7273 patch.

2.In arch/x86/include/asm/static_call.h, ARCH_DEFINE_STATIC_CALL_RET0_TRAMP()
is required for the RET0 implementation. And it is needed to solve the build
issues and therefore picked it from 5517d500829c68 patch.

3.The upstream changes for arch/x86/kernel/static_call.c have not been
backported due to the __static_call_transform() call, which introduces
additional dependencies and conflicts, and is not needed for the RET0
Implementation.

Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Signed-off-by: Frederic Weisbecker <frederic@kernel.org>
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Signed-off-by: Ingo Molnar <mingo@kernel.org>
Link: https://lkml.kernel.org/r/20210118141223.123667-2-frederic@kernel.org
Signed-off-by: Hemanth Selam <Hemanth.Selam@amd.com>