KVM: arm64: Fix PMU probe ordering
commit e840f42a49925707fca90e6c7a4095118fdb8c4d upstream

Russell reported that since 5.13, KVM's probing of the PMU has
started to fail on his HW. As it turns out, there is an implicit
ordering dependency between the architectural PMU probing code and
and KVM's own probing. If, due to probe ordering reasons, KVM probes
before the PMU driver, it will fail to detect the PMU and prevent it
from being advertised to guests as well as the VMM.

Obviously, this is one probing too many, and we should be able to
deal with any ordering.

Add a callback from the PMU code into KVM to advertise the registration
of a host CPU PMU, allowing for any probing order.

[Backport changes]

1.In arch/arm64/kvm/pmu-emul.c file, the 'int kvm_pmu_probe_pmuver()'
call is already changed to 'static int kvm_pmu_probe_pmuver()' call
in current kernel.

2.In include/kvm/arm_pmu.h file, the 'int kvm_pmu_probe_pmuver()' is not
present in current kernel.

Fixes: 5421db1be3b1 ("KVM: arm64: Divorce the perf code from oprofile helpers")
Reported-by: "Russell King (Oracle)" <linux@armlinux.org.uk>
Tested-by: Russell King (Oracle) <rmk+kernel@armlinux.org.uk>
Signed-off-by: Marc Zyngier <maz@kernel.org>
Link: https://lore.kernel.org/r/YUYRKVflRtUytzy5@shell.armlinux.org.uk
Cc: stable@vger.kernel.org
Signed-off-by: Hemanth Selam <Hemanth.Selam@amd.com>